move scripts into controllers/, dao/
fixed all references, hopefully no bugs anywhere
