+++ login.php
@@ -1,7 +1,5 @@
 <?php
 
-require("scripts/dao.php");
-
 /*=========================================================
  * Start the session and redirect if the user is logged in
  */
@@ -34,7 +32,7 @@ if (isset($_GET['feedback'])) { $feedback = $_GET['feedback']; $feedbackValid = 
 <body>
 
 	<div class="container">
-		<form class="form-signin" action="scripts/controller_login.php" method="post" role="form">
+		<form class="form-signin" action="controllers/controller_login.php" method="post" role="form">
 			<h1>Login</h1>
 <!--<img src="img/logo.png" class="img-responsive"> -->
 			<input type="hidden" name="controllerType" value="userLogin">
